Rename toon base entity module
