+++ homeassistant/components/toon/entity.py
